added 1 files
In this Project, We create  JavaScript Form Validation Project .This code snippet includes the HTML form structure with Bootstrap styling classes. The JavaScript code adds an event listener to the form's submit event and performs the necessary validations for each field based on the given criteria. If any validation criteria are not met, an error message is displayed next to the corresponding field. If there are no error messages, the form is considered valid and can be submitted. The clearErrors() function clears any previous error messages before performing the validations. The checkErrors() function checks if there are any error messages displayed, and if so, it returns true, indicating that the form should not be submitted.